feat: resolved all the changes of coderabit ai
