+++ frontend/ui/package.json
@@ -84,7 +84,6 @@
     "@tanstack/zod-form-adapter": "catalog:",
     "@tiptap/react": "catalog:",
     "@tiptap/starter-kit": "catalog:",
-    "apexcharts": "4.5.0",
     "aria-hidden": "1.2.4",
     "axios": "1.10.0",
     "chrono-node": "2.8.3",
@@ -102,7 +101,6 @@
     "motion": "12.7.4",
     "nanoid": "catalog:",
     "nuqs": "2.4.3",
-    "react-apexcharts": "1.7.0",
     "react-clock": "5.1.0",
     "react-day-picker": "9.7.0",
     "react-dropzone": "14.3.8",
