+++ frontend/ui/src/components/charts/index.ts
@@ -1,4 +1,3 @@
-export * from "./radial-bar-chart"
 export * from "./radial-task-status-chart"
 export * from "./spark-chart"
 export * from "./statisic-weekly-active-chart"
