+++ frontend/ui/src/components/charts/radial-task-status-chart.tsx
@@ -1,27 +1,11 @@
-import { Box, CardContainer } from "@incmix/ui"
+import { Box, Flex, Heading, Text } from "@incmix/ui"
 import { cn } from "@utils"
-import { lazy, useEffect, useState } from "react"
-import { useThemeStore,useAppearanceStore } from "@incmix/store/use-settings-store"
-import { Label, PolarAngleAxis, PolarRadiusAxis, RadialBar, RadialBarChart, ResponsiveContainer } from "recharts"
-// Dynamically import ApexCharts to avoid SSR issues
-import {
-  ChartConfig,
-  ChartContainer,
-  ChartTooltip,
-  ChartTooltipContent,
-} from "./shadcn-chart"
+import { PolarAngleAxis, RadialBar, RadialBarChart, ResponsiveContainer } from "recharts"
+
 export interface TaskItem {
-  /**
-   * Name of the task status
-   */
   name: string
-
-  /**
-   * Value/count of tasks in this status
-   */
   value: number
-
-
+  fill: string
 }
 
 interface RadialTaskStatusChartProps {
@@ -45,76 +29,29 @@ interface RadialTaskStatusChartProps {
   endAngle?: number
 
   /**
-   * Size of the hollow center as percentage
-   * @default "40%"
-   */
-  hollowSize?: string
-
-  /**
-   * Track background color
-   * @default var(--color-track)
-   */
-  trackBackground?: string
-
-  /**
    * Show labels below the chart
    * @default true
    */
   showLabels?: boolean
 }
-const data = [
-  {
-    name: "ongoing",
-    value: 45,
-    fill: "var(--indigo-9)",
-  },
-  {
-    name: "hold",
-    value: 55,
-    fill: "var(--orange-9)", 
-  },
-  {
-    name: "done",
-    value: 55,
-    fill: "var(--green-9)", 
-  },
-]
-export const description = "A radial chart with stacked sections"
-const chartData = [{ month: "january", ongoing: 1260, hold: 570, done: 570 }]
-const chartConfig = {
-  ongoing: {
-    label: "Ongoing",
-    color: "var(--indigo-9)",
-  },
-  hold: {
-    label: "Hold",
-    color: "var(--orange-9)",
-  },
-  done: {
-    label: "Done",
-    color: "var(--green-9)",
-  },
-} satisfies ChartConfig
+
+
 export function RadialTaskStatusChart({
-  // tasks = [
-  //   { name: "Ongoing", value: 420, color: 'pink' },
-  //   { name: "Hold", value: 210, color: 'yellow' },
-  //   { name: "Done", value: 200, color: 'purple' },
-  // ],
-  startAngle = -135,
-  endAngle = 135,
-  hollowSize = "40%",
+  tasks = [
+    { name: "Ongoing", value: 420,fill: "var(--indigo-9)" },
+    { name: "Hold", value: 210,fill: "var(--orange-9)" },
+    { name: "Done", value: 200,fill: "var(--green-9)" },
+  ],
+  startAngle = 180,
+  endAngle = 0,
   className,
-  trackBackground ,
 }: RadialTaskStatusChartProps) {
 
- 
-  const totalVisitors = chartData[0].ongoing + chartData[0].hold + chartData[0].done
-
+    const totalVisitors = tasks.reduce((sum, task) => sum + task.value, 0)
   return (
     <>
-      <div className="w-full relative h-72">
-      <div className="xl:h-[28rem] h-full ">
+      <Box className={cn("w-full relative h-72", className)}>
+      <Box className="xl:h-[28rem] h-full ">
         <ResponsiveContainer width="100%" height="100%">
           <RadialBarChart
             cx="50%"
@@ -122,21 +59,21 @@ export function RadialTaskStatusChart({
             innerRadius="60%"
             outerRadius="90%"
             barSize={20}
-            data={data}
-            startAngle={180}
-            endAngle={0}
+            data={tasks}
+            startAngle={startAngle}
+            endAngle={endAngle}
           >
             <PolarAngleAxis type="number" domain={[0, 100]} tick={false} tickLine={false} />
             <RadialBar dataKey="value" cornerRadius={4} fill="#f3f4f61d" background={{ fill: "#f3f4f61f" }} />
           </RadialBarChart>
         </ResponsiveContainer>
-      </div>
-      <div className="absolute inset-0 top-12 flex flex-col items-center justify-center">
-          <div className="text-4xl font-bold text-gray-12">1,830</div>
-          <div className="text-sm text-gray-11 mt-1">Visitors</div>
-        </div>
+      </Box>
+      <Flex justify="center" align="center" className="absolute inset-0 top-12 flex-col">
+          <Heading size="8">{totalVisitors}</Heading>
+          <Text className="text-gray-11 mt-1">Visitors</Text>
+        </Flex>
 
-    </div>
+    </Box>
 
     </>
   )
