+++ frontend/ui/src/components/dashboard/sidebar/widgets-data.tsx
@@ -6,7 +6,6 @@ import {
     ActiveTask,
     NewTasks,
     PostingTask,
-    ProjectWidgets2,
     StatisticWidgets2,
     TotalProject,
     TotalTasksChart,
@@ -23,6 +22,7 @@ import {
     LiveVisitors,
     InProgressTask,
     DoneTasks,
+    ProjectWidgets,
   } from "@incmix/ui/widgets";
 
 export const sidebarComponents = [
@@ -201,7 +201,7 @@ export const sidebarComponents = [
   },
   {
     slotId: "j",
-    component: <ProjectWidgets2 />,
+    component: <ProjectWidgets />,
     componentName: "project-widget",
     darkCompImage: dashboardImg.darkProjectChartImg,
     lightCompImage: dashboardImg.lightProjectChartImg,
