+++ frontend/ui/src/widgets/project/index.tsx
@@ -1,5 +1,4 @@
 export { ProjectWidgets } from "./project"
-export { ProjectWidgets2 } from "./project2"
 export { TotalProject } from "./total-project"
 export { ProjectTimelineWidgets } from "./project-timeline"
 export { ProjectListWidgets } from "./project-list"
